Background updates, bug fixes
- Updated the background image to use the large cloudy background
(clouds_highres.png). The older, smaller image (clouds.png) is still in
the files and can be switched by adjusting the downloaded image in
client.js and the Background image used in entity.js.
- Fix for the multi jump issue if not the host. The jumpPlayer method in
gameEngine now correctly checks if the players velocity is equal to zero
while also checking whether the player is allowed to multi-jump.
- Added a black border around Player text elements in the UI.